fix : use float

+++ ZooDeLille/app/src/main/java/com/example/zoodelille/data/api/object/ZooPosition.java
@@ -2,8 +2,8 @@ package com.example.zoodelille.data.api.object;
 
 public class ZooPosition {
     private int id;
-    private int longitude;
-    private int latitude;
+    private float longitude;
+    private float latitude;
 
     public ZooPosition() {
     }
@@ -21,19 +21,19 @@ public class ZooPosition {
         this.id = id;
     }
 
-    public int getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 }
